requirements.txt
flask==2.0.1
flask-login==0.5.0
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-wtf==1.0.0
flask-mail==0.9.1
flask-cors==3.0.10
email-validator==1.1.3
pillow==11.2.1
pandas==2.2.3
beautifulsoup4==4.10.0
python-dotenv==0.19.1
itsdangerous==2.0.1
werkzeug==2.0.2
Jinja2==3.0.2
requests==2.26.0 